+++ packages/extension/src/utils/getCookieKey.ts
@@ -13,13 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default as CirclePieChart } from './circlePieChart';
-export { default as Circle } from './circle';
-export { default as MatrixComponent } from './matrix/matrixComponent';
-export { default as MatrixComponentHorizontal } from './matrix/matrixComponent/matrixComponentHorizontal';
-export { default as Matrix } from './matrix';
-export { default as MessageBox } from './messageBox';
-export { default as Button } from './button';
-export { default as ProgressBar } from './progressBar';
+/**
+ * External dependencies.
+ */
+import { type Cookie as ParsedCookie } from 'simple-cookie';
 
-export type { MatrixComponentProps } from './matrix/matrixComponent';
+export const getCookieKey = (parsedCookie: ParsedCookie) => {
+  const cookieName = parsedCookie?.name;
+  const cookieDomain = parsedCookie?.domain;
+  const cookiePath = parsedCookie?.path;
+
+  return cookieName + cookieDomain + cookiePath;
+};