+++ packages/extension/src/view/settings/tests/index.tsx
@@ -1,0 +1,79 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * External dependencies.
+ */
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import { act } from 'react-dom/test-utils';
+import '@testing-library/jest-dom';
+import SinonChrome from 'sinon-chrome';
+
+/**
+ * Internal dependencies.
+ */
+import App from '../app';
+import { Provider as SettingsProvider } from '../stateProviders/syncSettingsStore';
+describe('Index', () => {
+  beforeAll(() => {
+    globalThis.chrome = SinonChrome as unknown as typeof chrome;
+    globalThis.chrome = {
+      ...SinonChrome,
+      storage: {
+        //@ts-ignore
+        local: {
+          clear: () => Promise.resolve(),
+          //@ts-ignore
+          onChanged: {
+            addListener: () => undefined,
+            removeListener: () => undefined,
+          },
+        },
+        sync: {
+          //@ts-ignore
+          // eslint-disable-next-line @typescript-eslint/no-unused-vars
+          get: (_, __) =>
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            new Promise<{ [key: string]: any }>((resolve) => {
+              resolve({
+                allowedNumberOfTabs: 'single',
+              });
+            }),
+          //@ts-ignore
+          onChanged: {
+            addListener: () => undefined,
+            removeListener: () => undefined,
+          },
+        },
+      },
+    };
+  });
+
+  it('All Providers should be added to DOM', async () => {
+    act(() =>
+      render(
+        <SettingsProvider>
+          <App />
+        </SettingsProvider>
+      )
+    );
+    expect(await screen.findByText('Settings')).toBeInTheDocument();
+  });
+
+  afterAll(() => {
+    globalThis.chrome = undefined as unknown as typeof chrome;
+  });
+});